+++ R/operations.R
@@ -34,6 +34,89 @@ as.global_operation <- function(x){
   structure(x, class = c("gce_global_operation", class(x)))
 }
 
+#' Deletes the specified Operations resource.
+#' 
+#' @seealso \href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
+#' 
+#' 
+#' @param operation Name of the Operations resource to delete
+#' @param project Project ID for this request
+#' @param zone Name of the zone for this request
+#' 
+#' @return TRUE if successful
+#' 
+#' @importFrom googleAuthR gar_api_generator
+#' @export
+gce_delete_op <- function(operation,
+                          project = gce_get_global_project(), 
+                          zone = gce_get_global_zone() ) {
+
+  
+  if(inherits(operation, c("gce_global_operation", "gce_zone_operation","gce_region_operation"))){
+    UseMethod("gce_delete_op", operation)
+  } else {
+    myMessage("No operation class found. Got: ", class(operation), level = 1)
+    return(operation)
+  }
+  
+}
+
+#' Deletes the specified zone-specific Operations resource.
+#' 
+#' @seealso \href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
+#' 
+#' 
+#' @param operation Name of the Operations resource to delete
+#' 
+#' @return The deleted operation
+#' 
+#' @importFrom googleAuthR gar_api_generator
+#' @export
+gce_delete_op.gce_zone_operation <- function(operation) {
+  
+  if(is.gce_zone_operation(operation)){
+    url <- operation$selfLink
+  } else {
+    stop("Not a gce_zone_operation")
+  }
+  
+  # compute.zoneOperations.delete
+  f <- gar_api_generator(url, "DELETE", data_parse_function = function(x) x)
+  
+  suppressWarnings(out <- f())
+  myMessage("Operation cancelled", level = 3)
+  as.zone_operation(out)
+  
+}
+
+#' Deletes the specified global Operations resource.
+#' 
+#' @seealso \href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
+#' 
+#' 
+#' @param operation Name of the Operations resource to delete
+#' 
+#' @return The deleted operation
+#' 
+#' @importFrom googleAuthR gar_api_generator
+#' @export
+gce_delete_op.gce_global_operation <- function(operation) {
+  
+  if(is.gce_global_operation(operation)){
+    url <- operation$selfLink
+  } else {
+    stop("Not a gce_global_operation")
+  }
+  
+  # compute.zoneOperations.delete
+  f <- gar_api_generator(url, "DELETE", data_parse_function = function(x) x)
+  
+  suppressWarnings(out <- f())
+  myMessage("Operation cancelled", level = 3)
+  as.zone_operation(out)
+  
+}
+
 #' Deletes the specified zone-specific Operations resource.
 #' 
 #' @seealso \href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
@@ -56,14 +139,9 @@ as.global_operation <- function(x){
 gce_delete_zone_op <- function(operation,
                                project = gce_get_global_project(), 
                                zone = gce_get_global_zone() ) {
-  url <- sprintf("https://www.googleapis.com/compute/v1/projects/%s/zones/%s/operations/%s", 
-                 project, zone, operation)
-  # compute.zoneOperations.delete
-  f <- gar_api_generator(url, "DELETE", data_parse_function = function(x) x)
   
-  suppressWarnings(out <- f())
-  myMessage("Operation cancelled", level = 3)
-  as.zone_operation(out)
+  .Deprecated("gce_delete_op", package = "googleComputeEngineR")
+  gce_delete_op(operation)
   
 }