+++ NEWS.md
@@ -11,6 +11,7 @@
 - Add global operation class
 - Add `open_webports` argument to `gce_vm` that will open web ports 80 and 443 if necessary
 - Add GPU functions
+- Migrate to use `system2` instead of `system` for cross-platform SSH (#35)
 
 # googleComputeEngineR 0.1.0