+++ NEWS.md
@@ -16,6 +16,7 @@
 - Add `gce_schedule_docker` and `gce_vm_scheduler` for easy Dockerfile scheduling
 - Add `gce_vm_logs` to quickly browse to an instance logs online
 - Fix custom machine types creation (#63) - thanks @Blaza
+- Set environment vars on VMs from metadata via `gce_metadata_env`
 
 # googleComputeEngineR 0.1.0